+++ docs/en_US/changelog.md
@@ -39,6 +39,10 @@ Fix June 08, 2020
 Add 15 octobre 2020
 ---
 
--   Adding gas temperature
+- Adding gas temperature
+- Adding power value
+- Adding ventilation value
+- Adding statistic
+Saving object is needed
 
 [Documentation](index.md)